java: loss of precission
Reviewed-on: https://gerrit.libreoffice.org/12000
Tested-by: Noel Grandin <noelgrandin@gmail.com>

+++ qadevOOo/runner/convwatch/BorderRemover.java
@@ -77,7 +77,7 @@ class BorderRemover
             }
 
             // calculate tolerance halve
-            double nTolerable = (_nTolerance * 256 / 100);
+            int nTolerable = (int) (_nTolerance * 256.0 / 100.0);
             if (nTolerable < 0)
             {
                 nTolerable = 0;
